feat: delete unsed aside schedule pop up files after refactory
