+++ web/app/components/ui/dialog.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 import * as DialogPrimitive from '@radix-ui/react-dialog';
 import { X } from 'lucide-react';
 
-import { cn } from '@/app/lib/utils';
+import { cn } from '@/app/utils/utils';
 
 const Dialog = DialogPrimitive.Root;
